Updated the deconvolution code to have the normalization constant equal to 1.
Originally, there was a crop on the FT of the PSF to update values smaller
than eps to be eps, but this then changed the physical interpretation of thePSF. It was now adding flux, when it should only be moving flux around.

By updating the code it renormalizes the PSF to be an integral of 1, thus
allowing the original interpretation of the PSF to be reinstated.