Added new script to calculate the analytical non/rotated 2D gaussians and their respective fourier transforms. Also, added two new notebooks titled "04x" that use the scripts and also use the functions in the scripts in the notebooks themselves for clarity.
NOTE: The reason for the disparity between the analytical and the FFT FTs was
due to the spacing between each frequency component in the analytical case
was too large, so it was giving too large of a value. Using
fft.fftfreq(input, d) and changing d <= 0.2 was useful in getting the answers
to be closer.